CDITCK-340 - Separate remote EJB tests from test archives
