+++ impl/src/main/java/org/jboss/cdi/tck/tests/extensions/producer/remote/ProducerProcessor.java
@@ -1,0 +1,62 @@
+/*
+ * JBoss, Home of Professional Open Source
+ * Copyright 2010, Red Hat, Inc., and individual contributors
+ * by the @authors tag. See the copyright.txt in the distribution for a
+ * full listing of individual contributors.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ * http://www.apache.org/licenses/LICENSE-2.0
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,  
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package org.jboss.cdi.tck.tests.extensions.producer.remote;
+
+import java.util.Set;
+
+import javax.enterprise.context.spi.CreationalContext;
+import javax.enterprise.event.Observes;
+import javax.enterprise.inject.spi.Extension;
+import javax.enterprise.inject.spi.InjectionPoint;
+import javax.enterprise.inject.spi.ProcessProducer;
+import javax.enterprise.inject.spi.ProcessProducerField;
+import javax.enterprise.inject.spi.Producer;
+
+public class ProducerProcessor implements Extension {
+
+    static boolean isOverriddenServiceProducerCalled = false;
+    static boolean isServiceProducerFieldObserved = false;
+
+    public void processServiceProducer(@Observes ProcessProducer<ServiceProducer, ServiceRemote> producerEvent) {
+        final Producer<ServiceRemote> producer = producerEvent.getProducer();
+        producerEvent.setProducer(new Producer<ServiceRemote>() {
+
+            public void dispose(ServiceRemote instance) {
+                producer.dispose(instance);
+            }
+
+            public Set<InjectionPoint> getInjectionPoints() {
+                return producer.getInjectionPoints();
+            }
+
+            public ServiceRemote produce(CreationalContext<ServiceRemote> ctx) {
+                isOverriddenServiceProducerCalled = true;
+                return producer.produce(ctx);
+            }
+        });
+    }
+
+    /**
+     * FIXME revise parameters order according to CDI-88/CDITCK-174 resolution
+     * 
+     * @param event
+     */
+    public void processServiceProducerField(@Observes ProcessProducerField<ServiceRemote, ServiceProducer> event) {
+        isServiceProducerFieldObserved = true;
+    }
+
+}
