+++ impl/src/main/java/org/jboss/cdi/tck/tests/extensions/producer/remote/RemoteProducerTest.java
@@ -1,0 +1,63 @@
+/*
+ * JBoss, Home of Professional Open Source
+ * Copyright 2010, Red Hat, Inc., and individual contributors
+ * by the @authors tag. See the copyright.txt in the distribution for a
+ * full listing of individual contributors.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ * http://www.apache.org/licenses/LICENSE-2.0
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package org.jboss.cdi.tck.tests.extensions.producer.remote;
+
+import static org.jboss.cdi.tck.TestGroups.JAVAEE_FULL;
+import static org.jboss.cdi.tck.cdi.Sections.PB;
+import static org.jboss.cdi.tck.cdi.Sections.PP;
+import static org.testng.Assert.assertEquals;
+import static org.testng.Assert.assertTrue;
+
+import org.jboss.arquillian.container.test.api.Deployment;
+import org.jboss.cdi.tck.AbstractTest;
+import org.jboss.cdi.tck.shrinkwrap.WebArchiveBuilder;
+import org.jboss.shrinkwrap.api.spec.WebArchive;
+import org.jboss.test.audit.annotations.SpecAssertion;
+import org.jboss.test.audit.annotations.SpecAssertions;
+import org.jboss.test.audit.annotations.SpecVersion;
+import org.testng.annotations.Test;
+
+/**
+ * Producer extension tests.
+ *
+ * @author David Allen
+ * @author Martin Kouba
+ */
+@Test(groups = JAVAEE_FULL) // The test archive contains a remote EJB
+@SpecVersion(spec = "cdi", version = "1.1 Final Release")
+public class RemoteProducerTest extends AbstractTest {
+
+    @Deployment
+    public static WebArchive createTestArchive() {
+        return new WebArchiveBuilder().withTestClassPackage(RemoteProducerTest.class).withExtension(ProducerProcessor.class).build();
+    }
+
+    @Test
+    @SpecAssertions({ @SpecAssertion(section = PP, id = "ac"), @SpecAssertion(section = PP, id = "e"),
+            @SpecAssertion(section = PP, id = "gb") })
+    public void testProduceCallsOverridenResourceProducerMethod() {
+        assertEquals(getContextualReference(ServiceRemote.class).ping(), "pong");
+        assertTrue(ProducerProcessor.isOverriddenServiceProducerCalled);
+    }
+
+    @Test
+    @SpecAssertions({ @SpecAssertion(section = PB, id = "ec") })
+    public void testResourceAreConsideredToBeProducerField() {
+        assertTrue(ProducerProcessor.isServiceProducerFieldObserved);
+    }
+
+}
