+++ impl/src/main/java/org/jboss/cdi/tck/tests/extensions/producer/remote/ServiceProducer.java
@@ -14,11 +14,14 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.jboss.cdi.tck.tests.implementation.enterprise.definition;
+package org.jboss.cdi.tck.tests.extensions.producer.remote;
 
-import javax.ejb.Remote;
+import javax.ejb.EJB;
+import javax.enterprise.inject.Produces;
 
-@Remote
-public interface DogRemote {
+public class ServiceProducer {
 
+    @Produces
+    @EJB
+    public ServiceRemote service;
 }
